Make disposeAnimations a transform
